requirements.txt
asgiref==3.4.1
awscli==1.16.252
certifi==2021.10.8
colorama==0.3.9
Django==3.2.9
kaggle==1.5.12
pyasn1==0.4.7
pygame==2.0.0
pytz==2021.3
readme-renderer==24.0
requests-toolbelt==0.9.1
rsa==3.4.2
sqlparse==0.4.2
twine==3.1.1
typing_extensions==4.0.0
mysqlclient
python-dotenv
djangorestframework
